refactor(micro-sam): split into CPU entry + GPU runtime (bump 0.3.0)
Adopt the model-runner architecture so training and inference can run at the
same time on separate GPUs:
- entry.py (EntryApp, CPU, 1 replica): Hypha/S3 transport, training data
  materialization, session orchestration, and routing to the runtime (composed
  by type hint). Returns the runtime's ndarray wire-dicts to the client.
- runtime.py (RuntimeApp, GPU, autoscale min=1/max=2, target=1): resident SAM
  encoder + ONE asyncio GPU lock over EVERY GPU op (serving AND training). A long
  training run holds one replica's lock, so a concurrent infer spins up and runs
  on the 2nd GPU replica.
- train_worker.py: fine-tuning runs in a subprocess (OS reclaims all training
  VRAM on exit); the resident inference model is evicted first so the subprocess
  owns the GPU — the 'evict when switching' requirement.
- training.py: run_training_blocking moved to train_worker.py; added
  write/read_training_params. requirements split entry/runtime; deployment.py removed.
